Refactor plan code into a new file (#656)
