+++ bodo/pandas/array_manager.py
@@ -242,7 +242,7 @@ class LazyArrayManager(ArrayManager, LazyMetadataMixin[ArrayManager]):
         return type(self)(arrays, new_axes, verify_integrity=False)
 
     def execute_plan(self):
-        from bodo.pandas.utils import execute_plan
+        from bodo.pandas.plan import execute_plan
 
         data = execute_plan(self._plan)
         if isinstance(data, BodoLazyWrapper):
@@ -482,7 +482,7 @@ class LazySingleArrayManager(SingleArrayManager, LazyMetadataMixin[SingleArrayMa
         return super().dtype
 
     def execute_plan(self):
-        from bodo.pandas.utils import execute_plan
+        from bodo.pandas.plan import execute_plan
 
         data = execute_plan(self._plan)
         if isinstance(data, BodoLazyWrapper):
